[RRTB-4] ignore cdk.out
