+++ .gitignore
@@ -4,6 +4,7 @@ Thumbs.db
 build/
 target/
 out/
+*/cdk.out/
 .micronaut/
 .idea
 *.iml
