+++ rrtb_daily_post_lambda/src/test/java/com/nb/SchedulerHandlerTest.java
@@ -1,12 +1,29 @@
 package com.nb;
 
+import org.assertj.core.api.Assertions;
+import org.junit.Ignore;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.utility.MountableFile;
+
+import java.io.File;
 import java.io.IOException;
 import java.util.concurrent.TimeUnit;
 
 class SchedulerHandlerTest extends AbstractTest {
 
+    @Test
+    @Ignore("the file does not exist as tests are running before projecg is packaged to jar")
+    void shouldCreateMountableFileforHostFileOnWindows() {
+        final MountableFile given = MountableFile.forHostPath("target/rrtb_daily_post_lambda-0.1.jar");
 
-    //    @Test
+        final String mountablePath = given.getResolvedPath();
+
+        final File actual = new File(mountablePath);
+        Assertions.assertThat(actual).exists();
+    }
+
+    @Test
+    @Ignore("the file does not exist as tests are running before projecg is packaged to jar")
     void should() throws IOException, InterruptedException {
         org.testcontainers.containers.Container.ExecResult execResult = LOCAL_STACK_CONTAINER.execInContainer(
                 "awslocal", "lambda", "create-function",
