--- rrtb_infra/cdk.out/MicronautAppStack.assets.json
@@ -1,58 +1,0 @@
-{
-  "version": "38.0.1",
-  "files": {
-    "4d226d8285c8415f85247b196ed465e1c95648487600a7efcc330efdf5062230": {
-      "source": {
-        "path": "asset.4d226d8285c8415f85247b196ed465e1c95648487600a7efcc330efdf5062230.jar",
-        "packaging": "file"
-      },
-      "destinations": {
-        "current_account-current_region": {
-          "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "4d226d8285c8415f85247b196ed465e1c95648487600a7efcc330efdf5062230.jar",
-          "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
-        }
-      }
-    },
-    "2819175352ad1ce0dae768e83fc328fb70fb5f10b4a8ff0ccbcb791f02b0716d": {
-      "source": {
-        "path": "asset.2819175352ad1ce0dae768e83fc328fb70fb5f10b4a8ff0ccbcb791f02b0716d",
-        "packaging": "zip"
-      },
-      "destinations": {
-        "current_account-current_region": {
-          "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "2819175352ad1ce0dae768e83fc328fb70fb5f10b4a8ff0ccbcb791f02b0716d.zip",
-          "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
-        }
-      }
-    },
-    "2fb0192b7cb277211376fbeb05ee84baa62e4092a97e2752612d507b3ada90fe": {
-      "source": {
-        "path": "asset.2fb0192b7cb277211376fbeb05ee84baa62e4092a97e2752612d507b3ada90fe.jar",
-        "packaging": "file"
-      },
-      "destinations": {
-        "current_account-current_region": {
-          "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "2fb0192b7cb277211376fbeb05ee84baa62e4092a97e2752612d507b3ada90fe.jar",
-          "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
-        }
-      }
-    },
-    "af8c4b276d8d29b0ab6ed14adc468cb2e3de1971da4ae8a5052ea268d67e6ff6": {
-      "source": {
-        "path": "MicronautAppStack.template.json",
-        "packaging": "file"
-      },
-      "destinations": {
-        "current_account-current_region": {
-          "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "af8c4b276d8d29b0ab6ed14adc468cb2e3de1971da4ae8a5052ea268d67e6ff6.json",
-          "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
-        }
-      }
-    }
-  },
-  "dockerImages": {}
-}
