--- rrtb_infra/cdk.out/cdk.out
@@ -1,1 +1,0 @@
-{"version":"38.0.1"}
