--- rrtb_infra/cdk.out/manifest.json
@@ -1,198 +1,0 @@
-{
-  "version": "38.0.1",
-  "artifacts": {
-    "RrtbAppStack.assets": {
-      "type": "cdk:asset-manifest",
-      "properties": {
-        "file": "RrtbAppStack.assets.json",
-        "requiresBootstrapStackVersion": 6,
-        "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version"
-      }
-    },
-    "RrtbAppStack": {
-      "type": "aws:cloudformation:stack",
-      "environment": "aws://unknown-account/unknown-region",
-      "properties": {
-        "templateFile": "RrtbAppStack.template.json",
-        "terminationProtection": false,
-        "validateOnSynth": false,
-        "notificationArns": [],
-        "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-deploy-role-${AWS::AccountId}-${AWS::Region}",
-        "cloudFormationExecutionRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-cfn-exec-role-${AWS::AccountId}-${AWS::Region}",
-        "stackTemplateAssetObjectUrl": "s3://cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}/e6b07770523f0c5eb430538d37c4dbfb4f8896a0b55d40e1370400b4cbb8f459.json",
-        "requiresBootstrapStackVersion": 6,
-        "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version",
-        "additionalDependencies": [
-          "RrtbAppStack.assets"
-        ],
-        "lookupRole": {
-          "arn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-lookup-role-${AWS::AccountId}-${AWS::Region}",
-          "requiresBootstrapStackVersion": 8,
-          "bootstrapStackVersionSsmParameter": "/cdk-bootstrap/hnb659fds/version"
-        }
-      },
-      "dependencies": [
-        "RrtbAppStack.assets"
-      ],
-      "metadata": {
-        "/RrtbAppStack/rrtb-posts/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbposts27CA6C36"
-          }
-        ],
-        "/RrtbAppStack/rrtb-posts/Policy/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbpostsPolicy10D5C8F2"
-          }
-        ],
-        "/RrtbAppStack/RrtbBucket": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "RrtbBucket"
-          }
-        ],
-        "/RrtbAppStack/rrtb-input-lambda/ServiceRole/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbinputlambdaServiceRoleD49D4211"
-          }
-        ],
-        "/RrtbAppStack/rrtb-input-lambda/ServiceRole/DefaultPolicy/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbinputlambdaServiceRoleDefaultPolicy7B29C601"
-          }
-        ],
-        "/RrtbAppStack/rrtb-input-lambda/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbinputlambda13AD63C7"
-          }
-        ],
-        "/RrtbAppStack/rrtb-input-lambda/LogRetention/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbinputlambdaLogRetention3301140A"
-          }
-        ],
-        "/RrtbAppStack/rrtb-input-lambda/FunctionUrl/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbinputlambdaFunctionUrl15A0D6B1"
-          }
-        ],
-        "/RrtbAppStack/rrtb-input-lambda/invoke-function-url": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbinputlambdainvokefunctionurlD415E684"
-          }
-        ],
-        "/RrtbAppStack/LogRetentionaae0aa3c5b4d4f87b02d85b201efdd8a/ServiceRole/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "LogRetentionaae0aa3c5b4d4f87b02d85b201efdd8aServiceRole9741ECFB"
-          }
-        ],
-        "/RrtbAppStack/LogRetentionaae0aa3c5b4d4f87b02d85b201efdd8a/ServiceRole/DefaultPolicy/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "LogRetentionaae0aa3c5b4d4f87b02d85b201efdd8aServiceRoleDefaultPolicyADDA7DEB"
-          }
-        ],
-        "/RrtbAppStack/LogRetentionaae0aa3c5b4d4f87b02d85b201efdd8a/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "LogRetentionaae0aa3c5b4d4f87b02d85b201efdd8aFD4BFC8A"
-          }
-        ],
-        "/RrtbAppStack/LatestNodeRuntimeMap": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "LatestNodeRuntimeMap"
-          }
-        ],
-        "/RrtbAppStack/RrtbApiUrl": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "RrtbApiUrl"
-          }
-        ],
-        "/RrtbAppStack/rrtb-daily-post-lambda/ServiceRole/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbdailypostlambdaServiceRoleD32021BE"
-          }
-        ],
-        "/RrtbAppStack/rrtb-daily-post-lambda/ServiceRole/DefaultPolicy/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbdailypostlambdaServiceRoleDefaultPolicy3D6AC866"
-          }
-        ],
-        "/RrtbAppStack/rrtb-daily-post-lambda/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbdailypostlambda50A850D9"
-          }
-        ],
-        "/RrtbAppStack/rrtb-daily-post-lambda/LogRetention/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbdailypostlambdaLogRetention3B2E18F9"
-          }
-        ],
-        "/RrtbAppStack/RrtbDailyPostLambda": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "RrtbDailyPostLambda"
-          }
-        ],
-        "/RrtbAppStack/rrtb-daily-rule/Resource": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbdailyruleE57E9924"
-          }
-        ],
-        "/RrtbAppStack/rrtb-daily-rule/AllowEventRuleRrtbAppStackrrtbdailypostlambdaF8BB91A7": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "rrtbdailyruleAllowEventRuleRrtbAppStackrrtbdailypostlambdaF8BB91A7B6E3DF5B"
-          }
-        ],
-        "/RrtbAppStack/CDKMetadata/Default": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "CDKMetadata"
-          }
-        ],
-        "/RrtbAppStack/CDKMetadata/Condition": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "CDKMetadataAvailable"
-          }
-        ],
-        "/RrtbAppStack/BootstrapVersion": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "BootstrapVersion"
-          }
-        ],
-        "/RrtbAppStack/CheckBootstrapVersion": [
-          {
-            "type": "aws:cdk:logicalId",
-            "data": "CheckBootstrapVersion"
-          }
-        ]
-      },
-      "displayName": "RrtbAppStack"
-    },
-    "Tree": {
-      "type": "cdk:tree",
-      "properties": {
-        "file": "tree.json"
-      }
-    }
-  }
-}
